fix: prove receivers per scope, and close what one file can prove
Review found a real false-positive vector: bindings were module-wide, so a
name proved in one function carried into any other function reusing the
spelling. A device pulled out of a dict was flagged because an unrelated
function above it had done a registry lookup into the same name. device is
one of the commonest locals in this ecosystem, so that is not a corner case.

Binding is per scope now, nested scopes inherit what encloses them, and a
parameter shadows whatever the enclosing scope proved. Three shapes were then
added back on evidence from core: comprehension generators, a registry kept
on an attribute (proved class-wide, since it is assigned in __init__ and read
elsewhere), and the registry's own devices/deleted_devices mappings.

Measured on home-assistant/core, 9 865 files and 5 963 textual
.config_entries occurrences: 51 findings, every one a device entry, 51 of the
54 present. The three misses are handed a device by another file, which no
single-file matcher can follow.

Test housekeeping in the same pass: test_local_scan.py covered two concerns
at 634 lines, so the local/index merge moved to test_local_merge.py.
sample_index had been copy-pasted into five files and FakeCoordinator into
two, with the two copies diverged; both now live in conftest.py once.
